Colocar los repositorios en un mismo archivo
