--- backend/app/database/repositories/course_repository.py
@@ -1,7 +1,0 @@
-from .sqlalchemy_repository import SQLAlchemyRepository
-from ...entities.course_entity import CourseEntity
-
-
-class CourseRepository(SQLAlchemyRepository):
-    def entity(self):
-        return CourseEntity
