--- backend/app/database/repositories/event_repository.py
@@ -1,7 +1,0 @@
-from .sqlalchemy_repository import SQLAlchemyRepository
-from ...entities.event_entity import EventEntity
-
-
-class EventRepository(SQLAlchemyRepository):
-    def entity(self):
-        return EventEntity
