--- backend/app/database/repositories/group_repository.py
@@ -1,7 +1,0 @@
-from .sqlalchemy_repository import SQLAlchemyRepository
-from ...entities.group_entity import GroupEntity
-
-
-class GroupRepository(SQLAlchemyRepository):
-    def entity(self):
-        return GroupEntity
