--- backend/app/database/repositories/local_repository.py
@@ -1,7 +1,0 @@
-from .sqlalchemy_repository import SQLAlchemyRepository
-from ...entities.local_entity import LocalEntity
-
-
-class LocalRepository(SQLAlchemyRepository):
-    def entity(self):
-        return LocalEntity
