--- backend/app/database/repositories/notification_repository.py
@@ -1,7 +1,0 @@
-from .sqlalchemy_repository import SQLAlchemyRepository
-from ...entities.notification_entity import NotificationEntity
-
-
-class NotificationRepository(SQLAlchemyRepository):
-    def entity(self):
-        return NotificationEntity
