--- backend/app/database/repositories/option_repository.py
@@ -1,7 +1,0 @@
-from .sqlalchemy_repository import SQLAlchemyRepository
-from ...entities.option_entity import OptionEntity
-
-
-class OptionRepository(SQLAlchemyRepository):
-    def entity(self):
-        return OptionEntity
