--- backend/app/database/repositories/permission_repository.py
@@ -1,7 +1,0 @@
-from .sqlalchemy_repository import SQLAlchemyRepository
-from ...entities.permission_entity import PermissionEntity
-
-
-class PermissionRepository(SQLAlchemyRepository):
-    def entity(self):
-        return PermissionEntity
