--- backend/app/database/repositories/resource_repository.py
@@ -1,7 +1,0 @@
-from .sqlalchemy_repository import SQLAlchemyRepository
-from ...entities.resource_entity import ResourceEntity
-
-
-class ResourceRepository(SQLAlchemyRepository):
-    def entity(self):
-        return ResourceEntity
