--- backend/app/database/repositories/role_repository.py
@@ -1,7 +1,0 @@
-from .sqlalchemy_repository import SQLAlchemyRepository
-from ...entities.role_entity import RoleEntity
-
-
-class RoleRepository(SQLAlchemyRepository):
-    def entity(self):
-        return RoleEntity
