--- backend/app/database/repositories/tag_repository.py
@@ -1,7 +1,0 @@
-from .sqlalchemy_repository import SQLAlchemyRepository
-from ...entities.tag_entity import TagEntity
-
-
-class TagRepository(SQLAlchemyRepository):
-    def entity(self):
-        return TagEntity
