--- backend/app/database/repositories/type_repository.py
@@ -1,7 +1,0 @@
-from .sqlalchemy_repository import SQLAlchemyRepository
-from ...entities.type_entity import TypeEntity
-
-
-class TypeRepository(SQLAlchemyRepository):
-    def entity(self):
-        return TypeEntity
