--- backend/app/database/repositories/user_repository.py
@@ -1,8 +1,0 @@
-from .sqlalchemy_repository import SQLAlchemyRepository
-from ...entities.user_entity import UserEntity
-
-
-class UserRepository(SQLAlchemyRepository):
-
-    def entity(self):
-        return UserEntity
