--- backend/app/database/repositories/votation_repository.py
@@ -1,7 +1,0 @@
-from .sqlalchemy_repository import SQLAlchemyRepository
-from ...entities.votation_entity import VotationEntity
-
-
-class VotationRepository(SQLAlchemyRepository):
-    def entity(self):
-        return VotationEntity
